feat: js file structure changed
