+++ src/main.js
@@ -1,14 +1,5 @@
 import getDataForGlobalTable from './globalTable';
 
-// Table Global Case
-const buttonSummaryAllPeriod = document.querySelector('.button__all-period');
-const buttonSummaryToday = document.querySelector('.button__last-day');
-// const buttonSummaryAbs = document.querySelector('.button__abs button');
-// const buttonSummaryPerPopulation = document.querySelector('.button__per-population');
-const totalConfirmed = document.querySelector('#total_confirm');
-const totalRecover = document.querySelector('#total_recover');
-const totalDeaths = document.querySelector('#total_deaths');
-
 // const url
 const urlSummary = 'https://api.covid19api.com/summary';
 
@@ -17,9 +8,9 @@ let data = '';
 
 let setDataToTableGlobal = '';
 
-async function getSummaryGlobalData(urlData) {
+async function getSummaryGlobalData() {
   console.log('function getSummaryGlobalData');
-  const url = urlData;
+  const url = urlSummary;
   const res = await fetch(url);
   data = await res.json();
   console.log('TotalConfirmed: ', data.Global.TotalConfirmed);
@@ -30,24 +21,13 @@ async function getSummaryGlobalData(urlData) {
 
 // first page load
 const firstPageLoad = () => {
-  getSummaryGlobalData(urlSummary);
+  getSummaryGlobalData();
 };
 
 setDataToTableGlobal = () => {
-  getDataForGlobalTable(totalConfirmed, totalRecover, totalDeaths, data);
+  getDataForGlobalTable(data);
 };
 
-// Listener
-buttonSummaryAllPeriod.addEventListener('click', () => {
-  buttonSummaryAllPeriod.classList.add('select');
-  buttonSummaryToday.classList.remove('select');
-  getSummaryGlobalData(urlSummary);
-});
-
-buttonSummaryToday.addEventListener('click', () => {
-  buttonSummaryAllPeriod.classList.remove('select');
-  buttonSummaryToday.classList.add('select');
-  getSummaryGlobalData(urlSummary);
-});
-
 firstPageLoad();
+
+export default getSummaryGlobalData;
